insulate postworld database functions with DB table check

+++ php/core/cache.php
@@ -79,6 +79,10 @@ function pw_clear_menu_caches(){
  * @param string $operator (Optional) Operator on which to query.
  */
 function pw_get_cache( $fields, $operator = 'AND' ){
+
+	if( !pw_config_in_db_tables('cache') )
+		return false;
+
 	global $wpdb;
 	$table_name = $wpdb->pw_prefix . 'cache';
 	$supported_fields = array( 'cache_name', 'cache_hash' );
@@ -142,6 +146,9 @@ function pw_get_cache( $fields, $operator = 'AND' ){
  */
 function pw_set_cache( $data ){
 
+	if( !pw_config_in_db_tables('cache') )
+		return false;
+
 	global $pw;
 	global $wpdb;
 	
@@ -189,6 +196,10 @@ function pw_set_cache( $data ){
 }
 
 function pw_delete_cache( $where ){
+
+	if( !pw_config_in_db_tables('cache') )
+		return false;
+
 	global $pw;
 	global $wpdb;
 	return $wpdb->delete(
@@ -198,6 +209,10 @@ function pw_delete_cache( $where ){
 }
 
 function pw_truncate_cache(){
+
+	if( !pw_config_in_db_tables('cache') )
+		return false;
+
 	global $pw;
 	global $wpdb;
 	return $wpdb->query("TRUNCATE TABLE `" . $wpdb->pw_prefix . "cache`");
@@ -227,6 +242,10 @@ function pw_get_cache_hash_content( $cache_hash ){
 
 
 function pw_get_cache_types_readout(){
+
+	if( !pw_config_in_db_tables('cache') )
+		return false;
+
 	// Returns an array containing the counts of each of the cache types
 	// ie. [{"cache_type":"feed","type_count":"4"},{"cache_type":"term-feed","type_count":"2"}]
 	global $wpdb;
@@ -612,6 +631,9 @@ function pw_cache_all_comment_points(){
 }
 
 function pw_clear_cron_logs ( $timestamp ){
+	if( !pw_config_in_db_tables('cron_logs') )
+		return false;
+
 	/*  • Count number of rows in wp_postworld_cron_logs (rows_before)
 		• Deletes all rows which are before the specified timestamp (rows_removed)
 		• Count number of rows after clearing (rows_after)
@@ -765,6 +787,9 @@ function pw_cache_shares ( $cache_all = FALSE ){
 
 
 function pw_get_most_recent_cache_shares_log(){
+	if( !pw_config_in_db_tables('cron_logs') )
+		return false;
+
 	global $wpdb;
 	$wpdb->show_errors();
 	$query="SELECT * FROM $wpdb->pw_prefix"."cron_logs  WHERE time_start = (SELECT MAX(time_start) FROM $wpdb->pw_prefix"."cron_logs where function_type = 'pw_cache_shares')";
@@ -773,6 +798,9 @@ function pw_get_most_recent_cache_shares_log(){
 }
 
 function pw_get_recent_shares_post_ids($last_time){
+	if( !pw_config_in_db_tables('shares') )
+		return false;
+
 	 global $wpdb;	
 	 $wpdb->show_errors();
 	 $query = "select DISTINCT  post_id from  $wpdb->pw_prefix"."shares where last_time>='$last_time'";
@@ -783,6 +811,9 @@ function pw_get_recent_shares_post_ids($last_time){
 }	
 
 function pw_get_recent_shares_author_ids($last_time){
+	if( !pw_config_in_db_tables('shares') )
+		return false;
+
 	 global $wpdb;	
 	 $wpdb->show_errors();
 	 $query = "select DISTINCT  author_id from  $wpdb->pw_prefix"."shares where last_time>='$last_time'";
@@ -791,6 +822,9 @@ function pw_get_recent_shares_author_ids($last_time){
 }	
 
 function pw_get_recent_shares_user_ids($last_time){
+	if( !pw_config_in_db_tables('shares') )
+		return false;
+
 	 global $wpdb;	
 	 $wpdb->show_errors();
 	 $query = "select DISTINCT user_id from  $wpdb->pw_prefix"."shares where last_time>='$last_time'";
@@ -822,6 +856,9 @@ function pw_get_all_user_ids_as_array(){
 
 //////////////// POST SHARES /////////////////////
 function pw_calculate_post_shares($post_id){
+	if( !pw_config_in_db_tables('shares') )
+		return false;
+
 	/*Calculates the total number of shares to the given post
 	Process
 	-Lookup the given post_id in the Shares table
@@ -840,6 +877,9 @@ function pw_calculate_post_shares($post_id){
 }
 
 function pw_cache_post_shares( $post_id ){
+
+	if( !pw_config_in_db_tables('post_meta') )
+		return false;
 
 	/*Caches the total number of shares to the given post
 	Process
@@ -911,6 +951,10 @@ function pw_calculate_user_shares( $user_id, $mode='both' ){
 }
 
 function pw_cache_user_shares( $user_id, $mode ){
+
+	if( !pw_config_in_db_tables('user_meta') )
+		return false;
+
 	/*
 	Caches the total number of shares relating to a given user
 	Process
@@ -954,6 +998,10 @@ function pw_cache_user_shares( $user_id, $mode ){
 }
 
 function pw_get_user_shares($user_id){
+
+	if( !pw_config_in_db_tables('user_meta') )
+		return false;
+
 	global $wpdb;
 	
 	if( pw_dev_mode() )
